Day 10

+++ 2023/src/day_template.rs
@@ -9,11 +9,11 @@ impl Data {
     }
 }
 
-fn part1(_data: Data) -> i32 {
+fn part1(_data: Data) -> i64 {
     return 0;
 }
 
-fn part2(_data: Data) -> i32 {
+fn part2(_data: Data) -> i64 {
     return 0;
 }
 
@@ -21,7 +21,6 @@ pub fn solve(contents: String) {
     let data = Data::from(contents.as_str());
 
     println!("Part 1: {}", part1(data.clone()));
-
     println!("Part 2: {}", part2(data));
 }
 
@@ -31,20 +30,20 @@ mod tests {
     use rstest::*;
 
     #[rstest]
-    #[case(include_str!("sample.txt"), Data {     })]
+    #[case(include_str!("sample.txt"), Data { })]
     fn test_parse_input(#[case] contents: &str, #[case] expected_value: Data) {
         assert_eq!(Data::from(contents), expected_value);
     }
 
     #[rstest]
-    #[case(include_str!("sample.txt"), 288)]
-    fn test_part1(#[case] contents: &str, #[case] expected_value: i32) {
+    #[case(include_str!("sample.txt"), 0)]
+    fn test_part1(#[case] contents: &str, #[case] expected_value: i64) {
         assert_eq!(part1(Data::from(contents)), expected_value);
     }
 
     #[rstest]
-    #[case(include_str!("sample.txt"), 71503)]
-    fn test_part2(#[case] contents: &str, #[case] expected_value: i32) {
+    #[case(include_str!("sample.txt"), 0)]
+    fn test_part2(#[case] contents: &str, #[case] expected_value: i64) {
         assert_eq!(part2(Data::from(contents)), expected_value);
     }
 }
